requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astrapy==0.7.4
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.39
botocore==1.34.39
cachetools==5.3.2
cassandra-driver==3.29.0
cassio==0.1.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dataclasses-json==0.6.4
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
emoji==2.10.1
extra-streamlit-components==0.1.60
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.2.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwt==1.3.1
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.3
langdetect==1.0.9
langsmith==0.0.90
llama-index==0.9.34
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
mistralai==0.0.12
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
orjson==3.9.13
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pypdf==4.0.1
python-dateutil==2.8.2
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
ragstack-ai==0.6.0
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.31.0
streamlit-modal==0.1.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.10.30
urllib3==2.0.7
validators==0.22.0
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
